LayoutTests/fast/shadow-dom/cannot-hit-test-into-shadow-media.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
<script>
window.jsTestIsAsync = true;

function runTest() {
    var bounds = videoElement.getBoundingClientRect();
    range = document.caretRangeFromPoint(bounds.left + 10, bounds.bottom - 10);
    shouldBe("range.commonAncestorContainer", "videoElement");
    finishJSTest();
}
</script>
</head>
<body onload="runTest()">
    <video id="videoElement" style="position:absolute; height: 240px; width: 320px;" controls></video>
</body>
</html>
