LayoutTests/fast/shadow-dom/cannot-hit-test-into-shadow-media-expected.txt

PASS range.commonAncestorContainer is videoElement
PASS successfullyParsed is true

TEST COMPLETE
